+++ demisto_sdk/commands/validate/file_validator.py
@@ -21,11 +21,12 @@ from demisto_sdk.commands.common.constants import (
     CHECKED_TYPES_REGEXES, CODE_FILES_REGEX, DIR_LIST_FOR_REGULAR_ENTETIES,
     IGNORED_TYPES_REGEXES, IMAGE_REGEX, INTEGRATION_REGEX, INTEGRATION_REGXES,
     JSON_ALL_DASHBOARDS_REGEXES, JSON_ALL_INCIDENT_TYPES_REGEXES,
-    JSON_ALL_LAYOUT_REGEXES, JSON_INDICATOR_AND_INCIDENT_FIELDS,
-    KNOWN_FILE_STATUSES, OLD_YML_FORMAT_FILE, PACKAGE_SCRIPTS_REGEXES,
+    JSON_ALL_INDICATOR_TYPES_REGEXES, JSON_ALL_LAYOUT_REGEXES,
+    JSON_INDICATOR_AND_INCIDENT_FIELDS, KNOWN_FILE_STATUSES,
+    OLD_YML_FORMAT_FILE, PACKAGE_SCRIPTS_REGEXES,
     PACKAGE_SUPPORTING_DIRECTORIES, PACKS_DIR, PACKS_DIRECTORIES,
-    PLAYBOOK_REGEX, PLAYBOOKS_REGEXES_LIST, REPUTATION_REGEX, SCHEMA_REGEX,
-    SCRIPT_REGEX, TEST_PLAYBOOK_REGEX, TEST_PLAYBOOKS_DIR, TESTS_DIRECTORIES,
+    PLAYBOOK_REGEX, PLAYBOOKS_REGEXES_LIST, SCHEMA_REGEX, SCRIPT_REGEX,
+    TEST_PLAYBOOK_REGEX, TEST_PLAYBOOKS_DIR, TESTS_DIRECTORIES,
     YML_ALL_SCRIPTS_REGEXES, YML_BETA_INTEGRATIONS_REGEXES,
     YML_INTEGRATION_REGEXES, Errors)
 from demisto_sdk.commands.common.hook_validations.conf_json import \
@@ -355,7 +356,7 @@ class FilesValidator:
                 if self.is_backward_check and not incident_field_validator.is_backward_compatible():
                     self._is_valid = False
 
-            elif checked_type(file_path, [REPUTATION_REGEX]):
+            elif checked_type(file_path, JSON_ALL_INDICATOR_TYPES_REGEXES):
                 reputation_validator = ReputationValidator(structure_validator)
                 if not reputation_validator.is_valid_file(validate_rn=True):
                     self._is_valid = False
@@ -463,7 +464,7 @@ class FilesValidator:
                 if not incident_field_validator.is_valid_file(validate_rn=False):
                     self._is_valid = False
 
-            elif checked_type(file_path, [REPUTATION_REGEX]) or file_type == 'reputation':
+            elif checked_type(file_path, JSON_ALL_INDICATOR_TYPES_REGEXES) or file_type == 'reputation':
                 reputation_validator = ReputationValidator(structure_validator)
                 if not reputation_validator.is_valid_file(validate_rn=False):
                     self._is_valid = False
@@ -615,7 +616,7 @@ class FilesValidator:
             if not incident_field_validator.is_valid_file(validate_rn=False):
                 self._is_valid = False
 
-        elif checked_type(file_path, [REPUTATION_REGEX]) or file_type == 'reputation':
+        elif checked_type(file_path, JSON_ALL_INDICATOR_TYPES_REGEXES) or file_type == 'reputation':
             reputation_validator = ReputationValidator(structure_validator)
             if not reputation_validator.is_valid_file(validate_rn=False):
                 self._is_valid = False