+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-
+* Added support for IndicatorTypes directory in packs (for `reputation` files, instead of Misc).
 
 #### 1.0.3
 * Fixed a bug in the **format** command where comments were being removed from YAML files.